helpers/dither/requirements.txt
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
hitherdither @ git+https://www.github.com/hbldh/hitherdither@b688094bb1e10bf5eda1ba999eb3c98a46ed9407
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.25.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
rich==13.4.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
